Initial setup to make flutter subtitle wrapper work better.
Added bloc and did some cleanup.
