+++ example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"video_player","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player-0.10.9+1/","dependencies":[]},{"name":"wakelock","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/wakelock-0.1.4+1/","dependencies":[]}],"android":[{"name":"video_player","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player-0.10.9+1/","dependencies":[]},{"name":"wakelock","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/wakelock-0.1.4+1/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"video_player_web","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player_web-0.1.2+3/","dependencies":[]}]},"dependencyGraph":[{"name":"video_player","dependencies":["video_player_web"]},{"name":"video_player_web","dependencies":[]},{"name":"wakelock","dependencies":[]}],"date_created":"2020-06-25 12:11:29.170164","version":"1.19.0-4.1.pre"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"video_player","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player-0.10.9+1/","dependencies":[]},{"name":"wakelock","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/wakelock-0.1.4+1/","dependencies":[]}],"android":[{"name":"video_player","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player-0.10.9+1/","dependencies":[]},{"name":"wakelock","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/wakelock-0.1.4+1/","dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[{"name":"video_player_web","path":"/Users/joran.dob/Flutter/flutter/.pub-cache/hosted/pub.dartlang.org/video_player_web-0.1.2+3/","dependencies":[]}]},"dependencyGraph":[{"name":"video_player","dependencies":["video_player_web"]},{"name":"video_player_web","dependencies":[]},{"name":"wakelock","dependencies":[]}],"date_created":"2020-06-25 12:52:02.340923","version":"1.19.0-4.1.pre"}
