+++ lib/bloc/subtitle_event.dart
@@ -1,0 +1,19 @@
+part of 'subtitle_bloc.dart';
+
+abstract class SubtitleEvent extends Equatable {
+  const SubtitleEvent();
+}
+
+class InitSubtitles extends SubtitleEvent {
+  final SubtitleController subtitleController;
+
+  InitSubtitles({@required this.subtitleController});
+
+  @override
+  List<Object> get props => [this.subtitleController];
+}
+
+class LoadSubtitle extends SubtitleEvent {
+  @override
+  List<Object> get props => [];
+}
