+++ lib/data/models/subtitles.dart
@@ -1,4 +1,4 @@
-import 'package:subtitle_wrapper_package/models/subtitle.dart';
+import 'package:subtitle_wrapper_package/data/models/subtitle.dart';
 
 class Subtitles {
   final List<Subtitle> subtitles;
