+++ lib/data/repository/subtitle_repository.dart
@@ -1,0 +1,89 @@
+import 'dart:convert';
+
+import 'package:flutter/foundation.dart';
+import 'package:subtitle_wrapper_package/data/models/subtitle.dart';
+import 'package:subtitle_wrapper_package/data/models/subtitles.dart';
+
+import 'package:http/http.dart' as http;
+import 'package:subtitle_wrapper_package/subtitle_controller.dart';
+
+abstract class SubtitleRepository {
+  Future<Subtitles> getSubtitles();
+}
+
+class SubtitleDataRepository extends SubtitleRepository {
+  final SubtitleController subtitleController;
+
+  SubtitleDataRepository({@required this.subtitleController});
+
+  @override
+  Future<Subtitles> getSubtitles() async {
+    String subtitlesContent = subtitleController.subtitlesContent;
+    String subtitleUrl = subtitleController.subtitleUrl;
+    SubtitleDecoder subtitleDecoder = subtitleController.subtitleDecoder;
+
+    RegExp regExp = new RegExp(
+      r"((\d{2}):(\d{2}):(\d{2})\.(\d+)) +--> +((\d{2}):(\d{2}):(\d{2})\.(\d{3})).*[\r\n]+\s*((?:(?!\r?\n\r?).)*(\r\n|\r|\n)(?:.*))",
+      caseSensitive: false,
+      multiLine: true,
+    );
+
+    if (subtitlesContent == null && subtitleUrl != null) {
+      http.Response response = await http.get(subtitleUrl);
+      if (response.statusCode == 200) {
+        subtitlesContent = subtitleDecoder == SubtitleDecoder.utf8
+            ? utf8.decode(
+                response.bodyBytes,
+                allowMalformed: true,
+              )
+            : latin1.decode(response.bodyBytes, allowInvalid: true);
+      }
+    }
+
+    List<RegExpMatch> matches = regExp.allMatches(subtitlesContent).toList();
+    List<Subtitle> subtitleList = List();
+
+    matches.forEach((RegExpMatch regExpMatch) {
+      int startTimeHours = int.parse(regExpMatch.group(2));
+      int startTimeMinutes = int.parse(regExpMatch.group(3));
+      int startTimeSeconds = int.parse(regExpMatch.group(4));
+      int startTimeMilliseconds = int.parse(regExpMatch.group(5));
+
+      int endTimeHours = int.parse(regExpMatch.group(7));
+      int endTimeMinutes = int.parse(regExpMatch.group(8));
+      int endTimeSeconds = int.parse(regExpMatch.group(9));
+      int endTimeMilliseconds = int.parse(regExpMatch.group(10));
+      String text = removeAllHtmlTags(regExpMatch.group(11));
+
+      Duration startTime = Duration(
+          hours: startTimeHours,
+          minutes: startTimeMinutes,
+          seconds: startTimeSeconds,
+          milliseconds: startTimeMilliseconds);
+      Duration endTime = Duration(
+          hours: endTimeHours,
+          minutes: endTimeMinutes,
+          seconds: endTimeSeconds,
+          milliseconds: endTimeMilliseconds);
+
+      subtitleList.add(
+          Subtitle(startTime: startTime, endTime: endTime, text: text.trim()));
+    });
+
+    Subtitles subtitles = Subtitles(subtitles: subtitleList);
+    return subtitles;
+  }
+
+  String removeAllHtmlTags(String htmlText) {
+    RegExp exp = RegExp(r"(<[^>]*>)", multiLine: true, caseSensitive: true);
+    String newHtmlText = htmlText;
+    exp.allMatches(htmlText).toList().forEach((RegExpMatch regExpMathc) {
+      if (regExpMathc.group(0) == "<br>") {
+        newHtmlText = newHtmlText.replaceAll(regExpMathc.group(0), '\n');
+      } else {
+        newHtmlText = newHtmlText.replaceAll(regExpMathc.group(0), '');
+      }
+    });
+    return newHtmlText;
+  }
+}
