+++ lib/subtitle_wrapper_package.dart
@@ -1,7 +1,10 @@
 library subtitle_wrapper_package;
 
 import 'package:flutter/material.dart';
-import 'package:subtitle_wrapper_package/models/style/subtitle_style.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:subtitle_wrapper_package/bloc/subtitle_bloc.dart';
+import 'package:subtitle_wrapper_package/data/models/style/subtitle_style.dart';
+import 'package:subtitle_wrapper_package/data/repository/subtitle_repository.dart';
 import 'package:subtitle_wrapper_package/subtitle_controller.dart';
 import 'package:subtitle_wrapper_package/subtitle_text_view.dart';
 import 'package:video_player/video_player.dart';
@@ -31,10 +34,20 @@ class SubTitleWrapper extends StatelessWidget {
                 bottom: subtitleStyle.position.bottom,
                 left: subtitleStyle.position.left,
                 right: subtitleStyle.position.right,
-                child: SubtitleTextView(
-                  subtitleController: subtitleController,
-                  videoPlayerController: videoPlayerController,
-                  subtitleStyle: subtitleStyle,
+                child: BlocProvider(
+                  create: (context) => SubtitleBloc(
+                    videoPlayerController: videoPlayerController,
+                    subtitleRepository: SubtitleDataRepository(
+                      subtitleController: subtitleController,
+                    ),
+                  )..add(
+                      InitSubtitles(
+                        subtitleController: subtitleController,
+                      ),
+                    ),
+                  child: SubtitleTextView(
+                    subtitleStyle: subtitleStyle,
+                  ),
                 ),
               )
             : Container(
