+++ pubspec.yaml
@@ -12,6 +12,9 @@ dependencies:
     sdk: flutter
   video_player: ^0.10.2+1
   http: ^0.12.0+2
+  bloc: ^4.0.0
+  flutter_bloc: ^4.0.0
+  equatable: ^1.2.0
 
 dev_dependencies:
   flutter_test:
